# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CameraControl API",
    "version": "1.0.0",
    "description": "Contrato vigente para api_version 1."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/esp32/status": {
      "get": {
        "operationId": "camera.esp32_status",
        "tags": [
          "esp32"
        ],
        "x-cameracontrol-feature": "esp32",
        "x-flask-endpoint": "camera.esp32_status",
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Esp32Status"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/timelapse/status": {
      "get": {
        "operationId": "timelapse.timelapse_status",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.timelapse_status",
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelapseStatus"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/timelapse/config": {
      "put": {
        "operationId": "timelapse.update_timelapse_config",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.update_timelapse_config",
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelapseStatus"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimelapseConfigRequest"
              }
            }
          }
        }
      }
    },
    "/api/timelapse/start": {
      "post": {
        "operationId": "timelapse.start_timelapse",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.start_timelapse",
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelapseStatus"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/timelapse/stop": {
      "post": {
        "operationId": "timelapse.stop_timelapse",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.stop_timelapse",
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelapseStatus"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/timelapse/folders": {
      "get": {
        "operationId": "timelapse.timelapse_folders",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.timelapse_folders",
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderList"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/timelapse/captures": {
      "get": {
        "operationId": "timelapse.timelapse_captures",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.timelapse_captures",
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureList"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "timelapse.delete_timelapse_captures",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.delete_timelapse_captures",
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureSelectionRequest"
              }
            }
          }
        }
      }
    },
    "/api/timelapse/folders/{folder_name}/download": {
      "get": {
        "operationId": "timelapse.download_timelapse_folder",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.download_timelapse_folder",
        "responses": {
          "200": {
            "description": "Respuesta correcta"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "folder_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/timelapse/capture/download": {
      "get": {
        "operationId": "timelapse.download_timelapse_capture",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.download_timelapse_capture",
        "responses": {
          "200": {
            "description": "Respuesta correcta"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/timelapse/capture/preview": {
      "get": {
        "operationId": "timelapse.preview_timelapse_capture",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.preview_timelapse_capture",
        "responses": {
          "200": {
            "description": "Respuesta correcta"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/timelapse/captures/download": {
      "post": {
        "operationId": "timelapse.download_selected_captures",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.download_selected_captures",
        "responses": {
          "200": {
            "description": "Respuesta correcta"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureSelectionRequest"
              }
            }
          }
        }
      }
    },
    "/api/timelapse/folders/{folder_name}": {
      "delete": {
        "operationId": "timelapse.delete_timelapse_folder",
        "tags": [
          "timelapse"
        ],
        "x-cameracontrol-feature": "timelapse",
        "x-flask-endpoint": "timelapse.delete_timelapse_folder",
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "folder_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "error"
        ]
      },
      "OkError": {
        "type": "object",
        "properties": {
          "ok": {
            "const": false
          },
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "ok",
          "error"
        ]
      },
      "StatusMessage": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": true,
        "required": [
          "status"
        ]
      },
...
      "Esp32Status": {
        "type": "object",
        "properties": {
          "connected": {
            "type": "boolean"
          },
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "device_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_state": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
...
      "DeleteResult": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "deleted": {
            "type": "integer"
          },
          "folder": {
            "type": "string"
          },
          "deleted_folder": {
            "type": "string"
          },
          "timelapse_resumed": {
            "type": "boolean"
          }
        },
        "additionalProperties": true,
        "required": [
          "ok"
        ]
      },
...
      "TimelapseConfigRequest": {
        "type": "object",
        "properties": {
          "interval_seconds": {
            "type": "integer",
            "minimum": 2
          },
          "width": {
            "type": "integer",
            "minimum": 1
          },
          "height": {
            "type": "integer",
            "minimum": 1
          },
          "auto_resume": {
            "type": "boolean"
          },
          "light_enabled": {
            "type": "boolean"
          },
          "light_intensity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "light_warmup_seconds": {
            "type": "integer",
            "minimum": 0,
            "maximum": 60
          },
          "folder_name": {
            "type": "string"
          },
          "save_sensor_readings": {
            "type": "boolean"
          },
          "capture_overlay_enabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "required": [
          "interval_seconds",
          "width",
          "height",
          "auto_resume"
        ]
      },
      "TimelapseStatus": {
        "type": "object",
        "properties": {},
        "additionalProperties": true
      },
      "CaptureSelectionRequest": {
        "type": "object",
        "properties": {
          "folder": {
            "type": "string"
          },
          "captures": {
            "type": "array",
            "minItems": 1,
            "maxItems": 5000,
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "folder",
          "captures"
        ]
      },
      "CaptureList": {
        "type": "object",
        "properties": {
          "folder": {
            "type": "string"
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "per_page": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "captures": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "total": {
            "type": "integer"
          },
          "pages": {
            "type": "integer",
            "minimum": 0
          }
        },
        "additionalProperties": false,
        "required": [
          "folder",
          "page",
          "per_page",
          "captures",
          "total",
          "pages"
        ]
      },
      "FolderList": {
        "type": "object",
        "properties": {
          "folders": {
            "type": "array",
            "items": {}
          },
          "selected": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "folders",
          "selected"
        ]
      },
...
    },
    "responses": {
      "Error": {
        "description": "Error según el envelope histórico del módulo",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "$ref": "#/components/schemas/OkError"
                },
                {
                  "$ref": "#/components/schemas/StatusMessage"
                }
              ]
            }
          }
        }
      }
    }
  }
}